flyttat logiken från ArticleController till ArticleService

+++ src/main/java/com/Molndal/WebShopService/Service/ArticleService.java
@@ -3,6 +3,9 @@ package com.Molndal.WebShopService.Service;
 import com.Molndal.WebShopService.Models.Article;
 import com.Molndal.WebShopService.Repository.ArticleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,11 +29,28 @@ public class ArticleService {
         return articleRepository.save(article);
     }
 
-    public Article updateArticle(Article article) {
-        return articleRepository.save(article);
+    //Bör lägga till felhantering i denna metod
+    public Article updateArticle(Long id, Article articleDetails) {
+
+        //Hämtar ut den befintliga artikeln
+        Article existingArticle = getOneArticle(id);
+
+        //Uppdaterar den befintliga artikeln
+        existingArticle.setName(articleDetails.getName());
+        existingArticle.setCost(articleDetails.getCost());
+        existingArticle.setDescription(articleDetails.getDescription());
+
+        //Returnerar och sparar den uppdaterade artikeln
+        return articleRepository.save(existingArticle);
     }
 
-    public void deleteOneArticle(Long id) {
-        articleRepository.deleteById(id);
+    public String deleteOneArticle(Long id) {
+        if (articleRepository.existsById(id)) {
+            articleRepository.deleteById(id);
+            return "Article with id: " + id + " was deleted";
+        } else {
+            return "Article with id: " + id + " does not exist";
+        }
     }
 }
+
